pyproject.toml
[tool.poetry]
authors = ["Mike Lissner <mike@free.law>"]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "License :: OSI Approved :: GNU Affero General Public License v3",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.13",
  "Intended Audience :: Developers",
  "Intended Audience :: Legal Industry",
]
description = "A free and powerful system for awareness and research of the American judicial system."
documentation = "https://github.com/freelawproject/courtlistener/wiki"
homepage = "https://www.courtlistener.com/"
include = ["manage.py"]
keywords = ["legal", "courts", "databases", "judiciary", "laws"] # Max: 5
license = "AGPL-3.0" # Should be spdx ID: https://spdx.org/licenses/
maintainers = ["Mike Lissner <mike@free.law>"]
name = "cl"
readme = "README.md"
repository = "https://github.com/freelawproject/courtlistener"
version = "0.14.24"

[tool.poetry.scripts]
cl-manage = "manage:main"

[tool.poetry.dependencies]
ada-url = "^1.21.0"
argparse = "*"
beautifulsoup4 = "==4.13.*"
celery = "^5.4.0"
certifi = "^2025.1.31"
courts-db = "*"
disposable-email-domains = "*"
Django = "^5.1.8"
django-cache-memoize = "==0.*"
django-cors-headers = "^4.7.0"
django-csp = "^3.8"
django-extensions = "^3.2.3"
django-filter = "^25.1"
django-localflavor = "^4.0"
django-markdown-deux = "^1.0.6"
django-mathfilters = "*"
django-pghistory = "^3.5.4"
django-ratelimit = "^4.1.0"
django-storages = "^1.14.5"
djangorestframework = "^3.14.0"
djangorestframework-xml = "^2.0.0"
feedparser = "^6.0.10"
httplib2 = "^0.22.0"
internetarchive = "^5.3.1"
ipaddress = "^1.0.16"
itypes = "^1.1.0"
kombu = "^5.5.1"
lxml = "^5.3.1"
markdown2 = "==2.5.*"
natsort = "^8.4.0"
ndg-httpsclient = "^0.5.1"
networkx = "^3.4.2"
nose = "*"
pandas = "^2.2.3"
pillow = "*"
pycparser = "^2.22"
pyopenssl = "*"
pyparsing = "^3.2.3"
python = ">=3.13, <3.14"
python-dateutil = "^2.9.0"
python-magic = "^0.4.27"
pytz = "*"
pyyaml = "^6.0.2"
redis = "^5.2.1"
requests = "^2.32.3"
simplejson = "^3.20.1"
timeout-decorator = "*"
unidecode = "*"
usaddress = "^0.5.13"
djangorestframework-filters = "1.0.0.dev2"
gunicorn = "^23.0.0"
django-hCaptcha = "^0.2.0"
reporters-db = "*"
django-waffle = "^4.2.0"
nameparser = "^1.1.3"
eyecite = "*"
scipy = "^1.15.2"
scikit-learn = "^1.6.1"
numpy = "^2.2.4"
datasketch = "^1.6.5"
PyStemmer = "^2.2.0.3"
factory-boy = "^3.3.3"
django-elasticsearch-dsl = "^8.0"
django-override-storage = "^0.3.2"
django-environ = "^0.12.0"
judge-pics = "^2.0.5"
django-admin-cursor-paginator = "^0.1.6"
sentry-sdk = {extras = ["celery", "django"], version = "^2.24.1"}
selenium = "^4.30.0"
ipython = "^9.0.2"
time-machine = "^2.16.0"
dateparser = "1.2.1"
types-dateparser = "^1.2.0.20250208"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
daphne = "^4.1.2"
httpx = {extras = ["http2"], version = "^0.28.1"}
django-model-utils = "^5.0.0"
django-permissions-policy = "^4.25.0"
tiktoken = "^0.9.0"
hyperscan = "^0.7.8"
openai = "^1.68.2"
seal-rookery = "^2.3.1"
types-pytz = "^2025.1.0.20250318"
psycopg = {version = "3.2.6", extras = ["binary", "pool"]}
boto3 = "^1.37.20"
django-tailwind = "^3.8.0"
drf-dynamic-fields = "^0.4.0"
django-ses = {extras = ["events"], version = "^4.4.0"}
juriscraper = "^2.6.26"
django-cotton = "^2.0.1"
drf-spectacular = "^0.28.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
types-redis = "^4.6.0.20241004"
pylint = "^3.3.6"
pytest = "^8.3.5"
pytest-django = "^4.10.0"
flake8 = "^7.1.2"
exrex = "^0.12.0"
tblib = "^3.0.0"
flynt = "^1.0.2"
mypy = "^1.15.0"
types-python-dateutil = "^2.9.0.20241206"
types-requests = "^2.32.0.20250306"
django-stubs = "^5.1.3"
djangorestframework-stubs = "^3.15.3"
black = "^25.1.0"
types-simplejson = "^3.20.0.20250318"
lxml-stubs = "^0.5.1"
django-debug-toolbar = "^5.1.0"
django-tailwind = {extras = ["reload"], version = "^3.8.0"}

[tool.black]
include = '''
(
    cl         # Just include a few directories for now
)*\.pyi?$
'''
exclude = '''
(
      \.eggs
    | .*\.egg-info
    | .*migrations.*
    | \.mypy_cache
    | __pycache__
    | \.venv
)
'''
line-length = 79

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 79
skip_glob = "*/migrations/*.py"

[tool.pylint.messages_control]
disable = "C0330, C0326"

[tool.pylint.format]
max-line-length = "79"

[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]
